Enable Unblind service to conditionally pass text to Talkback
Makes use of the accessibility event type TYPE_ANNOUNCEMENT for Unblind service to pass custom text to Talkback to read. This is done according to the Talkback UI element scrolling functionality, as the Unblind service responds to the TYPE_VIEW_ACCESSIBILITY_FOCUSED event, enabling good integration between the two services.